+++ index.html
@@ -26,5 +26,6 @@
         </div>
     </main>
 
-    <script language="javascript" type="text/javascript" src="js/main.js"></script>
+    <script language="javascript" type="text/javascript" src="js/calc.js"></script>
+    <script language="javascript" type="text/javascript" src="js/dom.js"></script>
 </body>
