+++ js/calc.js
@@ -1,0 +1,91 @@
+function getDirection(directAngle) {
+    //Umrechnung von einem Winkel in den dazugehörigen Vektor
+    let x;
+    let y;
+    if (directAngle > 360) {
+        directAngle = directAngle - 360;
+    }
+    if (directAngle < 0) {
+        directAngle = 360 + directAngle;
+    }
+    let angle = (Math.PI / 180) * directAngle;
+
+    if (directAngle <= 90) {
+        let xa = Math.sin(angle);
+        let ya = Math.cos(angle);
+
+        x = xa / (xa + ya);
+        y = ya / (xa + ya);
+    } else if (directAngle <= 180) {
+        let xa = Math.sin(angle);
+        let ya = Math.cos(angle);
+
+        x = xa / (xa + Math.abs(ya));
+        y = ya / (xa + Math.abs(ya));
+    } else if (directAngle <= 270) {
+        let xa = Math.sin(angle);
+        let ya = Math.cos(angle);
+
+        x = xa / Math.abs(xa + ya);
+        y = ya / Math.abs(xa + ya);
+    } else if (directAngle <= 360) {
+        let xa = Math.sin(angle);
+        let ya = Math.cos(angle);
+
+        x = xa / (Math.abs(xa) + ya);
+        y = ya / (Math.abs(xa) + ya);
+    }
+    return {
+        x: round(x),
+        y: round(y),
+    }
+}
+
+function getAngle(x, y) {
+    //Umrechnung von einem Vektor in den dazugehörigen Winkel
+    if ((x === 0 && y === 0) || (x === 1 && y === 1)) {
+        //keine Bewegung === keine Richung
+        return;
+    }
+    //wir können nicht durch 0 teilen, aber durch annähernd 0
+    if (y === 0) {
+        y = 0.00000001;
+    }
+    if (x === 0) {
+        x = 0.00000001;
+    }
+
+    let alpha = Math.round((180 / Math.PI) * Math.atan(x / y));
+    let angle;
+
+    //arctan hat nur einen Ergebnisbereich von 0 - 90° -> es müssen noch Anpassungen unternommen werden
+    if (x > 0 && y > 0) {
+        angle = alpha;
+    } else if (y < 0) {
+        angle = alpha + 180;
+    } else if (x < 0) {
+        angle = alpha + 360;
+    }
+
+    return angle;
+}
+
+function round(n) {
+    // 2 decimal places
+    let number = (Math.round(n * 100) / 100)
+    return number;
+}
+
+if (module !== undefined) {
+    module.exports = {
+        getDirection,
+        getAngle,
+        round
+    }
+}
+
+if (window) {
+    window.getDirection = getDirection
+    window.getAngle = getAngle
+    window.round = round 
+}
